--- common/table_info.go
@@ -1,692 +1,0 @@
-package common
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/pingcap/log"
-	"github.com/pingcap/tidb/pkg/parser/model"
-	"github.com/pingcap/tidb/pkg/parser/mysql"
-	"github.com/pingcap/tidb/pkg/table/tables"
-	"github.com/pingcap/tidb/pkg/types"
-	datumTypes "github.com/pingcap/tidb/pkg/types"
-	"github.com/pingcap/tidb/pkg/util/rowcodec"
-	"github.com/pingcap/tiflow/pkg/util"
-	"go.uber.org/zap"
-)
-
-// ColumnFlagType is for encapsulating the flag operations for different flags.
-type ColumnFlagType util.Flag
-
-const (
-	// BinaryFlag means the column charset is binary
-	BinaryFlag ColumnFlagType = 1 << ColumnFlagType(iota)
-	// HandleKeyFlag means the column is selected as the handle key
-	// The handleKey is chosen by the following rules in the order:
-	// 1. if the table has primary key, it's the handle key.
-	// 2. If the table has not null unique key, it's the handle key.
-	// 3. If the table has no primary key and no not null unique key, it has no handleKey.
-	HandleKeyFlag
-	// GeneratedColumnFlag means the column is a generated column
-	GeneratedColumnFlag
-	// PrimaryKeyFlag means the column is primary key
-	PrimaryKeyFlag
-	// UniqueKeyFlag means the column is unique key
-	UniqueKeyFlag
-	// MultipleKeyFlag means the column is multiple key
-	MultipleKeyFlag
-	// NullableFlag means the column is nullable
-	NullableFlag
-	// UnsignedFlag means the column stores an unsigned integer
-	UnsignedFlag
-)
-
-// SetIsBinary sets BinaryFlag
-func (b *ColumnFlagType) SetIsBinary() {
-	(*Flag)(b).Add(Flag(BinaryFlag))
-}
-
-// UnsetIsBinary unsets BinaryFlag
-func (b *ColumnFlagType) UnsetIsBinary() {
-	(*Flag)(b).Remove(Flag(BinaryFlag))
-}
-
-// IsBinary shows whether BinaryFlag is set
-func (b *ColumnFlagType) IsBinary() bool {
-	return (*Flag)(b).HasAll(Flag(BinaryFlag))
-}
-
-// SetIsHandleKey sets HandleKey
-func (b *ColumnFlagType) SetIsHandleKey() {
-	(*Flag)(b).Add(Flag(HandleKeyFlag))
-}
-
-// UnsetIsHandleKey unsets HandleKey
-func (b *ColumnFlagType) UnsetIsHandleKey() {
-	(*Flag)(b).Remove(Flag(HandleKeyFlag))
-}
-
-// IsHandleKey shows whether HandleKey is set
-func (b *ColumnFlagType) IsHandleKey() bool {
-	return (*Flag)(b).HasAll(Flag(HandleKeyFlag))
-}
-
-// SetIsGeneratedColumn sets GeneratedColumn
-func (b *ColumnFlagType) SetIsGeneratedColumn() {
-	(*Flag)(b).Add(Flag(GeneratedColumnFlag))
-}
-
-// UnsetIsGeneratedColumn unsets GeneratedColumn
-func (b *ColumnFlagType) UnsetIsGeneratedColumn() {
-	(*Flag)(b).Remove(Flag(GeneratedColumnFlag))
-}
-
-// IsGeneratedColumn shows whether GeneratedColumn is set
-func (b *ColumnFlagType) IsGeneratedColumn() bool {
-	return (*Flag)(b).HasAll(Flag(GeneratedColumnFlag))
-}
-
-// SetIsPrimaryKey sets PrimaryKeyFlag
-func (b *ColumnFlagType) SetIsPrimaryKey() {
-	(*Flag)(b).Add(Flag(PrimaryKeyFlag))
-}
-
-// UnsetIsPrimaryKey unsets PrimaryKeyFlag
-func (b *ColumnFlagType) UnsetIsPrimaryKey() {
-	(*Flag)(b).Remove(Flag(PrimaryKeyFlag))
-}
-
-// IsPrimaryKey shows whether PrimaryKeyFlag is set
-func (b *ColumnFlagType) IsPrimaryKey() bool {
-	return (*Flag)(b).HasAll(Flag(PrimaryKeyFlag))
-}
-
-// SetIsUniqueKey sets UniqueKeyFlag
-func (b *ColumnFlagType) SetIsUniqueKey() {
-	(*Flag)(b).Add(Flag(UniqueKeyFlag))
-}
-
-// UnsetIsUniqueKey unsets UniqueKeyFlag
-func (b *ColumnFlagType) UnsetIsUniqueKey() {
-	(*Flag)(b).Remove(Flag(UniqueKeyFlag))
-}
-
-// IsUniqueKey shows whether UniqueKeyFlag is set
-func (b *ColumnFlagType) IsUniqueKey() bool {
-	return (*Flag)(b).HasAll(Flag(UniqueKeyFlag))
-}
-
-// IsMultipleKey shows whether MultipleKeyFlag is set
-func (b *ColumnFlagType) IsMultipleKey() bool {
-	return (*Flag)(b).HasAll(Flag(MultipleKeyFlag))
-}
-
-// SetIsMultipleKey sets MultipleKeyFlag
-func (b *ColumnFlagType) SetIsMultipleKey() {
-	(*Flag)(b).Add(Flag(MultipleKeyFlag))
-}
-
-// UnsetIsMultipleKey unsets MultipleKeyFlag
-func (b *ColumnFlagType) UnsetIsMultipleKey() {
-	(*Flag)(b).Remove(Flag(MultipleKeyFlag))
-}
-
-// IsNullable shows whether NullableFlag is set
-func (b *ColumnFlagType) IsNullable() bool {
-	return (*Flag)(b).HasAll(Flag(NullableFlag))
-}
-
-// SetIsNullable sets NullableFlag
-func (b *ColumnFlagType) SetIsNullable() {
-	(*Flag)(b).Add(Flag(NullableFlag))
-}
-
-// UnsetIsNullable unsets NullableFlag
-func (b *ColumnFlagType) UnsetIsNullable() {
-	(*Flag)(b).Remove(Flag(NullableFlag))
-}
-
-// IsUnsigned shows whether UnsignedFlag is set
-func (b *ColumnFlagType) IsUnsigned() bool {
-	return (*Flag)(b).HasAll(Flag(UnsignedFlag))
-}
-
-// SetIsUnsigned sets UnsignedFlag
-func (b *ColumnFlagType) SetIsUnsigned() {
-	(*Flag)(b).Add(Flag(UnsignedFlag))
-}
-
-// UnsetIsUnsigned unsets UnsignedFlag
-func (b *ColumnFlagType) UnsetIsUnsigned() {
-	(*Flag)(b).Remove(Flag(UnsignedFlag))
-}
-
-// TableName represents name of a table, includes table name and schema name.
-type TableName struct {
-	Schema      string `toml:"db-name" msg:"db-name"`
-	Table       string `toml:"tbl-name" msg:"tbl-name"`
-	TableID     int64  `toml:"tbl-id" msg:"tbl-id"`
-	IsPartition bool   `toml:"is-partition" msg:"is-partition"`
-}
-
-// String implements fmt.Stringer interface.
-func (t TableName) String() string {
-	return fmt.Sprintf("%s.%s", t.Schema, t.Table)
-}
-
-// QuoteSchema quotes a full table name
-func QuoteSchema(schema string, table string) string {
-	return fmt.Sprintf("`%s`.`%s`", EscapeName(schema), EscapeName(table))
-}
-
-// QuoteName wraps a name with "`"
-func QuoteName(name string) string {
-	return "`" + EscapeName(name) + "`"
-}
-
-// EscapeName replaces all "`" in name with double "`"
-func EscapeName(name string) string {
-	return strings.Replace(name, "`", "``", -1)
-}
-
-// QuoteString returns quoted full table name
-func (t TableName) QuoteString() string {
-	return QuoteSchema(t.Schema, t.Table)
-}
-
-// GetSchema returns schema name.
-func (t *TableName) GetSchema() string {
-	return t.Schema
-}
-
-// GetTable returns table name.
-func (t *TableName) GetTable() string {
-	return t.Table
-}
-
-// GetTableID returns table ID.
-func (t *TableName) GetTableID() int64 {
-	return t.TableID
-}
-
-const (
-	// HandleIndexPKIsHandle represents that the handle index is the pk and the pk is the handle
-	HandleIndexPKIsHandle = -1
-	// HandleIndexTableIneligible represents that the table is ineligible
-	HandleIndexTableIneligible = -2
-)
-
-// TableInfo provides meta data describing a DB table.
-type TableInfo struct {
-	*model.TableInfo
-	SchemaID int64
-	// NOTICE: We probably store the logical ID inside TableName,
-	// not the physical ID.
-	// For normal table, there is only one ID, which is the physical ID.
-	// AKA TIDB_TABLE_ID.
-	// For partitioned table, there are two kinds of ID:
-	// 1. TIDB_PARTITION_ID is the physical ID of the partition.
-	// 2. TIDB_TABLE_ID is the logical ID of the table.
-	// In general, we always use the physical ID to represent a table, but we
-	// record the logical ID from the DDL event(job.BinlogInfo.TableInfo).
-	// So be careful when using the TableInfo.
-	TableName TableName
-	// Version record the tso of create the table info.
-	Version uint64
-	// ColumnID -> offset in model.TableInfo.Columns
-	columnsOffset map[int64]int
-	// Column name -> ColumnID
-	nameToColID map[string]int64
-
-	hasUniqueColumn bool
-
-	// ColumnID -> offset in RowChangedEvents.Columns.
-	RowColumnsOffset map[int64]int
-
-	ColumnsFlag map[int64]*ColumnFlagType
-
-	// the mounter will choose this index to output delete events
-	// special value:
-	// HandleIndexPKIsHandle(-1) : pk is handle
-	// HandleIndexTableIneligible(-2) : the table is not eligible
-	HandleIndexID int64
-
-	// IndexColumnsOffset store the offset of the columns in row changed events for
-	// unique index and primary key
-	// The reason why we need this is that the Indexes in TableInfo
-	// will not contain the PK if it is create in statement like:
-	// create table t (a int primary key, b int unique key);
-	// Every element in first dimension is a index, and the second dimension is the columns offset
-	// for example:
-	// table has 3 columns: a, b, c
-	// pk: a
-	// index1: a, b
-	// index2: a, c
-	// indexColumnsOffset: [[0], [0, 1], [0, 2]]
-	IndexColumnsOffset [][]int
-
-	// The following 3 fields, should only be used to decode datum from the raw value bytes, do not abuse those field.
-	// rowColInfos extend the model.ColumnInfo with some extra information
-	// it's the same length and order with the model.TableInfo.Columns
-	rowColInfos    []rowcodec.ColInfo
-	rowColFieldTps map[int64]*types.FieldType
-	// only for new row format decoder
-	handleColID []int64
-
-	// number of virtual columns
-	virtualColumnCount int
-	// rowColInfosWithoutVirtualCols is the same as rowColInfos, but without virtual columns
-	rowColInfosWithoutVirtualCols *[]rowcodec.ColInfo
-}
-
-func (ti *TableInfo) initRowColInfosWithoutVirtualCols() {
-	if ti.virtualColumnCount == 0 {
-		ti.rowColInfosWithoutVirtualCols = &ti.rowColInfos
-		return
-	}
-	colInfos := make([]rowcodec.ColInfo, 0, len(ti.rowColInfos)-ti.virtualColumnCount)
-	for i, col := range ti.Columns {
-		if IsColCDCVisible(col) {
-			colInfos = append(colInfos, ti.rowColInfos[i])
-		}
-	}
-	if len(colInfos) != len(ti.rowColInfos)-ti.virtualColumnCount {
-		log.Panic("invalid rowColInfosWithoutVirtualCols",
-			zap.Int("len(colInfos)", len(colInfos)),
-			zap.Int("len(ti.rowColInfos)", len(ti.rowColInfos)),
-			zap.Int("ti.virtualColumnCount", ti.virtualColumnCount))
-	}
-	ti.rowColInfosWithoutVirtualCols = &colInfos
-}
-
-func (ti *TableInfo) findHandleIndex() {
-	if ti.HandleIndexID == HandleIndexPKIsHandle {
-		// pk is handle
-		return
-	}
-	handleIndexOffset := -1
-	for i, idx := range ti.Indices {
-		if !ti.IsIndexUnique(idx) {
-			continue
-		}
-		if idx.Primary {
-			handleIndexOffset = i
-			break
-		}
-		if handleIndexOffset < 0 {
-			handleIndexOffset = i
-		} else {
-			if len(ti.Indices[handleIndexOffset].Columns) > len(ti.Indices[i].Columns) ||
-				(len(ti.Indices[handleIndexOffset].Columns) == len(ti.Indices[i].Columns) &&
-					ti.Indices[handleIndexOffset].ID > ti.Indices[i].ID) {
-				handleIndexOffset = i
-			}
-		}
-	}
-	if handleIndexOffset >= 0 {
-		ti.HandleIndexID = ti.Indices[handleIndexOffset].ID
-	}
-}
-
-func (ti *TableInfo) initColumnsFlag() {
-	for _, colInfo := range ti.Columns {
-		var flag ColumnFlagType
-		if colInfo.GetCharset() == "binary" {
-			flag.SetIsBinary()
-		}
-		if colInfo.IsGenerated() {
-			flag.SetIsGeneratedColumn()
-		}
-		if mysql.HasPriKeyFlag(colInfo.GetFlag()) {
-			flag.SetIsPrimaryKey()
-			if ti.HandleIndexID == HandleIndexPKIsHandle {
-				flag.SetIsHandleKey()
-			}
-		}
-		if mysql.HasUniKeyFlag(colInfo.GetFlag()) {
-			flag.SetIsUniqueKey()
-		}
-		if !mysql.HasNotNullFlag(colInfo.GetFlag()) {
-			flag.SetIsNullable()
-		}
-		if mysql.HasMultipleKeyFlag(colInfo.GetFlag()) {
-			flag.SetIsMultipleKey()
-		}
-		if mysql.HasUnsignedFlag(colInfo.GetFlag()) {
-			flag.SetIsUnsigned()
-		}
-		ti.ColumnsFlag[colInfo.ID] = &flag
-	}
-
-	// In TiDB, just as in MySQL, only the first column of an index can be marked as "multiple key" or "unique key",
-	// and only the first column of a unique index may be marked as "unique key".
-	// See https://dev.mysql.com/doc/refman/5.7/en/show-columns.html.
-	// Yet if an index has multiple columns, we would like to easily determine that all those columns are indexed,
-	// which is crucial for the completeness of the information we pass to the downstream.
-	// Therefore, instead of using the MySQL standard,
-	// we made our own decision to mark all columns in an index with the appropriate flag(s).
-	for _, idxInfo := range ti.Indices {
-		for _, idxCol := range idxInfo.Columns {
-			colInfo := ti.Columns[idxCol.Offset]
-			flag := ti.ColumnsFlag[colInfo.ID]
-			if idxInfo.Primary {
-				flag.SetIsPrimaryKey()
-			} else if idxInfo.Unique {
-				flag.SetIsUniqueKey()
-			}
-			if len(idxInfo.Columns) > 1 {
-				flag.SetIsMultipleKey()
-			}
-			if idxInfo.ID == ti.HandleIndexID && ti.HandleIndexID >= 0 {
-				flag.SetIsHandleKey()
-			}
-			ti.ColumnsFlag[colInfo.ID] = flag
-		}
-	}
-}
-
-// GetColumnInfo returns the column info by ID
-func (ti *TableInfo) GetColumnInfo(colID int64) (info *model.ColumnInfo, exist bool) {
-	colOffset, exist := ti.columnsOffset[colID]
-	if !exist {
-		return nil, false
-	}
-	return ti.Columns[colOffset], true
-}
-
-// ForceGetColumnInfo return the column info by ID
-// Caller must ensure `colID` exists
-func (ti *TableInfo) ForceGetColumnInfo(colID int64) *model.ColumnInfo {
-	colInfo, ok := ti.GetColumnInfo(colID)
-	if !ok {
-		log.Panic("invalid column id", zap.Int64("columnID", colID))
-	}
-	return colInfo
-}
-
-// ForceGetColumnFlagType return the column flag type by ID
-// Caller must ensure `colID` exists
-func (ti *TableInfo) ForceGetColumnFlagType(colID int64) *ColumnFlagType {
-	flag, ok := ti.ColumnsFlag[colID]
-	if !ok {
-		log.Panic("invalid column id", zap.Int64("columnID", colID))
-	}
-	return flag
-}
-
-// ForceGetColumnName return the column name by ID
-// Caller must ensure `colID` exists
-func (ti *TableInfo) ForceGetColumnName(colID int64) string {
-	return ti.ForceGetColumnInfo(colID).Name.O
-}
-
-// ForceGetColumnIDByName return column ID by column name
-// Caller must ensure `colID` exists
-func (ti *TableInfo) ForceGetColumnIDByName(name string) int64 {
-	colID, ok := ti.nameToColID[name]
-	if !ok {
-		log.Panic("invalid column name", zap.String("column", name))
-	}
-	return colID
-}
-
-// GetSchemaName returns the schema name of the table
-func (ti *TableInfo) GetSchemaName() string {
-	return ti.TableName.Schema
-}
-
-// GetTableName returns the table name of the table
-func (ti *TableInfo) GetTableName() string {
-	return ti.TableName.Table
-}
-
-// GetSchemaNamePtr returns the pointer to the schema name of the table
-func (ti *TableInfo) GetSchemaNamePtr() *string {
-	return &ti.TableName.Schema
-}
-
-// GetTableNamePtr returns the pointer to the table name of the table
-func (ti *TableInfo) GetTableNamePtr() *string {
-	return &ti.TableName.Table
-}
-
-// IsPartitionTable returns whether the table is partition table
-func (ti *TableInfo) IsPartitionTable() bool {
-	return ti.TableName.IsPartition
-}
-
-func (ti *TableInfo) String() string {
-	return fmt.Sprintf("TableInfo, ID: %d, Name:%s, ColNum: %d, IdxNum: %d, PKIsHandle: %t", ti.ID, ti.TableName, len(ti.Columns), len(ti.Indices), ti.PKIsHandle)
-}
-
-// GetRowColInfos returns all column infos for rowcodec
-func (ti *TableInfo) GetRowColInfos() ([]int64, map[int64]*types.FieldType, []rowcodec.ColInfo) {
-	return ti.handleColID, ti.rowColFieldTps, ti.rowColInfos
-}
-
-// GetColInfosForRowChangedEvent return column infos for non-virtual columns
-// The column order in the result is the same as the order in its corresponding RowChangedEvent
-func (ti *TableInfo) GetColInfosForRowChangedEvent() []rowcodec.ColInfo {
-	return *ti.rowColInfosWithoutVirtualCols
-}
-
-// IsColCDCVisible returns whether the col is visible for CDC
-func IsColCDCVisible(col *model.ColumnInfo) bool {
-	// this column is a virtual generated column
-	if col.IsGenerated() && !col.GeneratedStored {
-		return false
-	}
-	return true
-}
-
-// HasUniqueColumn returns whether the table has a unique column
-func (ti *TableInfo) HasUniqueColumn() bool {
-	return ti.hasUniqueColumn
-}
-
-// HasVirtualColumns returns whether the table has virtual columns
-func (ti *TableInfo) HasVirtualColumns() bool {
-	return ti.virtualColumnCount > 0
-}
-
-// IsEligible returns whether the table is a eligible table
-func (ti *TableInfo) IsEligible(forceReplicate bool) bool {
-	// Sequence is not supported yet, TiCDC needs to filter all sequence tables.
-	// See https://github.com/pingcap/tiflow/issues/4559
-	if ti.IsSequence() {
-		return false
-	}
-	if forceReplicate {
-		return true
-	}
-	if ti.IsView() {
-		return true
-	}
-	return ti.HasUniqueColumn()
-}
-
-// IsIndexUnique returns whether the index is unique
-func (ti *TableInfo) IsIndexUnique(indexInfo *model.IndexInfo) bool {
-	if indexInfo.Primary {
-		return true
-	}
-	if indexInfo.Unique {
-		for _, col := range indexInfo.Columns {
-			colInfo := ti.Columns[col.Offset]
-			if !mysql.HasNotNullFlag(colInfo.GetFlag()) {
-				return false
-			}
-			// this column is a virtual generated column
-			if colInfo.IsGenerated() && !colInfo.GeneratedStored {
-				return false
-			}
-		}
-		return true
-	}
-	return false
-}
-
-// Clone clones the TableInfo
-func (ti *TableInfo) Clone() *TableInfo {
-	return WrapTableInfo(ti.SchemaID, ti.TableName.Schema, ti.Version, ti.TableInfo.Clone())
-}
-
-// GetIndex return the corresponding index by the given name.
-func (ti *TableInfo) GetIndex(name string) *model.IndexInfo {
-	for _, index := range ti.Indices {
-		if index != nil && index.Name.O == name {
-			return index
-		}
-	}
-	return nil
-}
-
-// IndexByName returns the index columns and offsets of the corresponding index by name
-func (ti *TableInfo) IndexByName(name string) ([]string, []int, bool) {
-	index := ti.GetIndex(name)
-	if index == nil {
-		return nil, nil, false
-	}
-	names := make([]string, 0, len(index.Columns))
-	offset := make([]int, 0, len(index.Columns))
-	for _, col := range index.Columns {
-		names = append(names, col.Name.O)
-		offset = append(offset, col.Offset)
-	}
-	return names, offset, true
-}
-
-// OffsetsByNames returns the column offsets of the corresponding columns by names
-// If any column does not exist, return false
-func (ti *TableInfo) OffsetsByNames(names []string) ([]int, bool) {
-	// todo: optimize it
-	columnOffsets := make(map[string]int, len(ti.Columns))
-	for _, col := range ti.Columns {
-		if col != nil {
-			columnOffsets[col.Name.O] = col.Offset
-		}
-	}
-
-	result := make([]int, 0, len(names))
-	for _, col := range names {
-		offset, ok := columnOffsets[col]
-		if !ok {
-			return nil, false
-		}
-		result = append(result, offset)
-	}
-
-	return result, true
-}
-
-// GetPrimaryKeyColumnNames returns the primary key column names
-func (ti *TableInfo) GetPrimaryKeyColumnNames() []string {
-	var result []string
-	if ti.PKIsHandle {
-		result = append(result, ti.GetPkColInfo().Name.O)
-		return result
-	}
-
-	indexInfo := ti.GetPrimaryKey()
-	if indexInfo != nil {
-		for _, col := range indexInfo.Columns {
-			result = append(result, col.Name.O)
-		}
-	}
-	return result
-}
-
-// WrapTableInfo creates a TableInfo from a model.TableInfo
-func WrapTableInfo(schemaID int64, schemaName string, version uint64, info *model.TableInfo) *TableInfo {
-	ti := &TableInfo{
-		TableInfo: info,
-		SchemaID:  schemaID,
-		TableName: TableName{
-			Schema:      schemaName,
-			Table:       info.Name.O,
-			TableID:     info.ID,
-			IsPartition: info.GetPartitionInfo() != nil,
-		},
-		hasUniqueColumn:  false,
-		Version:          version,
-		columnsOffset:    make(map[int64]int, len(info.Columns)),
-		nameToColID:      make(map[string]int64, len(info.Columns)),
-		RowColumnsOffset: make(map[int64]int, len(info.Columns)),
-		ColumnsFlag:      make(map[int64]*ColumnFlagType, len(info.Columns)),
-		handleColID:      []int64{-1},
-		HandleIndexID:    HandleIndexTableIneligible,
-		rowColInfos:      make([]rowcodec.ColInfo, len(info.Columns)),
-		rowColFieldTps:   make(map[int64]*types.FieldType, len(info.Columns)),
-	}
-
-	rowColumnsCurrentOffset := 0
-
-	ti.virtualColumnCount = 0
-	for i, col := range ti.Columns {
-		ti.columnsOffset[col.ID] = i
-		pkIsHandle := false
-		if IsColCDCVisible(col) {
-			ti.nameToColID[col.Name.O] = col.ID
-			ti.RowColumnsOffset[col.ID] = rowColumnsCurrentOffset
-			rowColumnsCurrentOffset++
-			pkIsHandle = (ti.PKIsHandle && mysql.HasPriKeyFlag(col.GetFlag())) || col.ID == model.ExtraHandleID
-			if pkIsHandle {
-				// pk is handle
-				ti.handleColID = []int64{col.ID}
-				ti.HandleIndexID = HandleIndexPKIsHandle
-				ti.hasUniqueColumn = true
-				ti.IndexColumnsOffset = append(ti.IndexColumnsOffset, []int{ti.RowColumnsOffset[col.ID]})
-			} else if ti.IsCommonHandle {
-				ti.HandleIndexID = HandleIndexPKIsHandle
-				ti.handleColID = ti.handleColID[:0]
-				pkIdx := tables.FindPrimaryIndex(info)
-				for _, pkCol := range pkIdx.Columns {
-					id := info.Columns[pkCol.Offset].ID
-					ti.handleColID = append(ti.handleColID, id)
-				}
-			}
-		} else {
-			ti.virtualColumnCount += 1
-		}
-		ti.rowColInfos[i] = rowcodec.ColInfo{
-			ID:            col.ID,
-			IsPKHandle:    pkIsHandle,
-			Ft:            col.FieldType.Clone(),
-			VirtualGenCol: col.IsGenerated(),
-		}
-		ti.rowColFieldTps[col.ID] = ti.rowColInfos[i].Ft
-	}
-
-	for _, idx := range ti.Indices {
-		if ti.IsIndexUnique(idx) {
-			ti.hasUniqueColumn = true
-		}
-		if idx.Primary || idx.Unique {
-			indexColOffset := make([]int, 0, len(idx.Columns))
-			for _, idxCol := range idx.Columns {
-				colInfo := ti.Columns[idxCol.Offset]
-				if IsColCDCVisible(colInfo) {
-					indexColOffset = append(indexColOffset, ti.RowColumnsOffset[colInfo.ID])
-				}
-			}
-			if len(indexColOffset) > 0 {
-				ti.IndexColumnsOffset = append(ti.IndexColumnsOffset, indexColOffset)
-			}
-		}
-	}
-
-	ti.initRowColInfosWithoutVirtualCols()
-	ti.findHandleIndex()
-	ti.initColumnsFlag()
-	return ti
-}
-
-// GetColumnDefaultValue returns the default definition of a column.
-func GetColumnDefaultValue(col *model.ColumnInfo) interface{} {
-	defaultValue := col.GetDefaultValue()
-	if defaultValue == nil {
-		defaultValue = col.GetOriginDefaultValue()
-	}
-	defaultDatum := datumTypes.NewDatum(defaultValue)
-	return defaultDatum.GetValue()
-}
