--- common/types.go
@@ -1,15 +1,0 @@
-package common
-
-import (
-	"encoding/binary"
-
-	"github.com/google/uuid"
-)
-
-type Timestamp uint64
-type TableID int64
-type DispatcherID uuid.UUID
-type DatabaseID int64
-type SchemaID int64
-
-var DefaultEndian = binary.LittleEndian
