+++ downstreamadapter/dispatcher/event_dispatcher_task.go
@@ -16,7 +16,7 @@ package dispatcher
 import (
 	"time"
 
-	"github.com/flowbehappy/tigate/common"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/flowbehappy/tigate/utils/threadpool"
 )
