+++ downstreamadapter/dispatcher/table_event_dispatcher.go
@@ -16,8 +16,8 @@ package dispatcher
 import (
 	"time"
 
-	"github.com/flowbehappy/tigate/common"
 	"github.com/flowbehappy/tigate/downstreamadapter/sink"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/flowbehappy/tigate/utils/threadpool"
 	"github.com/google/uuid"
 )
