+++ downstreamadapter/dispatchermanager/heartbeat_task.go
@@ -16,8 +16,8 @@ package dispatchermanager
 import (
 	"time"
 
-	"github.com/flowbehappy/tigate/common"
 	"github.com/flowbehappy/tigate/downstreamadapter/dispatcher"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/flowbehappy/tigate/utils/threadpool"
 	"github.com/google/uuid"
 	"github.com/ngaut/log"
