+++ downstreamadapter/sink/conflictdetector/conflict_detector.go
@@ -14,8 +14,8 @@
 package conflictdetector
 
 import (
-	"github.com/flowbehappy/tigate/common"
 	"github.com/flowbehappy/tigate/downstreamadapter/sink/types"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/flowbehappy/tigate/utils/threadpool"
 	"github.com/pingcap/log"
 	"go.uber.org/atomic"
