+++ downstreamadapter/sink/conflictdetector/txn_cache.go
@@ -16,7 +16,7 @@ package conflictdetector
 import (
 	"sync/atomic"
 
-	"github.com/flowbehappy/tigate/common"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/pingcap/log"
 	"go.uber.org/zap"
 )
