+++ downstreamadapter/writer/dml_helper_test.go
@@ -3,7 +3,7 @@ package writer
 import (
 	"testing"
 
-	"github.com/flowbehappy/tigate/common"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/pingcap/tidb/pkg/parser/charset"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/stretchr/testify/require"
