+++ logservice/schemastore/schemastore.go
@@ -6,7 +6,7 @@ import (
 	"math"
 	"sync"
 
-	"github.com/flowbehappy/tigate/common"
+	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/parser/model"
